# src/swagger.yaml
swagger: '2.0'
info:
  description: ''
  version: 1.0.0
  title: API
host: 'localhost:3000'
basePath: /
tags:
  - name: User
schemes:
  - http
paths:

  /users:
    post:
      tags:
        - User
      summary: Adicionar um novo usuário
      description: ''
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/User'
      responses:
        '404':
          description: Invalid input

    get:
      tags:
        - User
      summary: Recupera todos os usuários
      description: ''
      operationId: findUser
      produces:
        - application/json
      responses:
        '200':
          description: successful operation
          schema:
            type: "array"
            items:
              $ref: "#/definitions/User"
        '404':
          description: User not found

  '/users/{userId}':
    put:
      tags:
        - User
      summary: Update an existing department
      description: ''
      operationId: updateDepartment
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: departmentId
          in: path
          description: ID of user to return
          required: true
          type: integer
          format: int64
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/User'
      responses:
        '404':
          description: User not found
    get:
      tags:
        - User
      summary: Find User by ID
      description: Returns a single User
      produces:
        - application/json
      parameters:
        - name: userId
          in: path
          description: ID of User to return
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: successful operation
          schema:
            $ref: '#/definitions/User'
        '404':
          description: User not found
    delete:
      tags:
        - User
      summary: Deletes a User
      description: ''
      produces:
        - application/json
      parameters:
        - name: userId
          in: path
          description: User id to delete
          required: true
          type: integer
          format: int64
      responses:
        '404':
          description: User not found

definitions:

  User:
    type: object
    properties:
      id:
        type: integer
        format: int64
      username:
        type: string
      password:
        type: string
      active:
        type: boolean